Handling search\filter buttons inside the Menu component instead of Root.
